+++ Frontend/pet-app/src/components/login/form.tsx
@@ -1,4 +1,4 @@
-import Input from "@/components/ui/input";
+import Input from "@/components/login/input";
 
 const Form = () => {
   return (
